requirements.txt
roboflow
opencv-python
pillow
matplotlib
numpy